refactor(ui): Move item count functions to helpers
Counting items like vulnerabilities will be needed in several places, so
move the item counting functions from inside product statistics cards
to helpers.
